backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
beautifulsoup4>=4.12.0
pypdf>=5.0.0
groq>=0.11.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
# Heavy-duty ingestion (no API key required for Crawl4AI browser path)
crawl4ai>=0.4.0
docling>=2.0.0
playwright>=1.40.0
